Huge refactor to make mpu6050 work
I2C lib was changed, by some reason if we reuse the opened bus
to write to some register, this set have no effect.

Also add a new message to protocol to send debug messages.

Changed double vars to float to reduce program size.

+++ src/mpu6050.h
@@ -3,7 +3,7 @@
 
 int mpu6050_init(void);
 void mpu6050_calibrate(void);
-void mpu6050_gyro_get(double *x, double *y, double *z);
-void mpu6050_accel_get(double *x, double *y, double *z);
+void mpu6050_gyro_get(float *x, float *y, float *z);
+void mpu6050_accel_get(float *x, float *y, float *z);
 
 #endif
